started with comment lines p.s. just started

+++ C/Assignment1/Calculator/mul.c
@@ -1,5 +1,5 @@
 #include<stdio.h>
-long int mul(int a, int b)
+long long int mul(int a, int b)
 {
 	return a * b;
 }
